+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Fix an issue where failing to load Datadog modules on Web threw an error (and potentially broke application loading).
 * Add the ability to specify a sampling rate for loggers.
 * Add a "NoOp" platform, usable when performing headless Flutter widget tests.